support cleaned trash

+++ seahub/tickets/tickets.py
@@ -28,7 +28,7 @@ from seahub.project.seadb_api import SeaDBAPI
 from seahub.tickets.ticket_utils import get_ticket, get_ticket_comments, \
     check_ticket_comment_creation_interval, get_ticket_comment_by_pk, check_ticket_creation_interval, \
     convert_ticket_select_column_name_to_option_id, TABLE_TICKETS, get_tickets_by_ids, get_my_tickets, \
-    delete_ticket_comments_by_ids
+    delete_ticket_comments_by_ids, get_deleted_tickets_ids
 
 SEAQA_VERSION = getattr(settings, 'SEAQA_VERSION', 'Dev')
 
@@ -314,8 +314,6 @@ class TicketsAPIView(APIView):
                 updated_row[TicketsTable.tags.name] = row_data.get('tags')
             if 'type' in row_data:
                 updated_row[TicketsTable.type.name] = row_data.get('type')
-            if 'deleted' in row_data:
-                updated_row[TicketsTable.deleted.name] = bool(row_data.get('deleted'))
             if 'content' in row_data:
                 content_dict = row_data.get('content')
                 if not isinstance(content_dict, dict):
@@ -531,7 +529,6 @@ class TicketAPIView(APIView):
             return api_error(status.HTTP_403_FORBIDDEN, error_msg)
 
         ticket_state_name = request.data.get('state')
-        deleted = request.data.get('deleted', '')
 
         is_update_type = 'type' in request.data
         type_name = request.data.get('type') or None
@@ -617,8 +614,6 @@ class TicketAPIView(APIView):
             participants = ticket.get('participants') or []
             if username not in participants:
                 participants.append(username)
-            if deleted is not None:
-                update_row[TicketsTable.deleted.name] = bool(deleted)
 
             now_datetime = datetime.datetime.now(datetime.UTC).isoformat()
             if ticket_state_name or ticket_state_name == '':
@@ -1256,10 +1251,10 @@ class TicketTrashAPIView(APIView):
             return api_error(status.HTTP_500_INTERNAL_SERVER_ERROR, error_msg)
         return Response({'tickets': tickets, 'columns': columns})
 
-    def delete(self, request, project_uuid):
+    def put(self, request, project_uuid):
         ticket_ids = request.data.get('ticket_ids')
         if not ticket_ids:
-            error_msg = 'ticket_ids is required.'
+            error_msg = 'tickets_data is required.'
             return api_error(status.HTTP_400_BAD_REQUEST, error_msg)
 
         if not is_org_context(request):
@@ -1279,22 +1274,54 @@ class TicketTrashAPIView(APIView):
             error_msg = 'Permission denied.'
             return api_error(status.HTTP_403_FORBIDDEN, error_msg)
 
+        seadb_api = SeaDBAPI(username)
+        update_rows = []
+        now_datetime = datetime.datetime.now(datetime.UTC).isoformat()
+        for ticket_id in ticket_ids:
+            updated_row = {
+                TicketsTable.modified_time.name: now_datetime,
+                TicketsTable.deleted.name: False,
+            }
+            update_rows.append(
+                {
+                    'pk': int(ticket_id),
+                    'row': updated_row,
+                }
+            )
+
+        if update_rows:
+            try:
+                seadb_api.update_rows(project_uuid, 'tickets', update_rows)
+            except Exception as e:
+                logger.exception(e)
+                error_msg = 'Internal Server Error'
+                return api_error(status.HTTP_500_INTERNAL_SERVER_ERROR, error_msg)
+
+        return Response({'success': True})
+
+    def delete(self, request, project_uuid):
+        if not is_org_context(request):
+            error_msg = 'Feature is not enabled.'
+            return api_error(status.HTTP_403_FORBIDDEN, error_msg)
+
+        # resource check
+        project = Projects.objects.get_project_by_uuid(project_uuid)
+        if not project:
+            error_msg = 'Project not found.'
+            return api_error(status.HTTP_404_NOT_FOUND, error_msg)
+        workspace = project.workspace
+
+        # permission check
+        username = request.user.username
+        if not check_project_permission(username, workspace.owner):
+            error_msg = 'Permission denied.'
+            return api_error(status.HTTP_403_FORBIDDEN, error_msg)
+
         try:
             seadb_api = SeaDBAPI(username)
-        except Exception as e:
-            logger.error(e)
-            error_msg = 'Internal Server Error'
-            return api_error(status.HTTP_500_INTERNAL_SERVER_ERROR, error_msg)
-
-        tickets = get_tickets_by_ids(seadb_api, project_uuid, ticket_ids)
-        exist_ticket_ids = [ticket.get('_pk') for ticket in tickets]
-        fail_ticket_ids = []
-        for ticket_id in ticket_ids:
-            if int(ticket_id) not in exist_ticket_ids:
-                fail_ticket_ids.append(int(ticket_id))
-
-        need_delete_ticket_ids = [int(ticket_id) for ticket_id in set(ticket_ids) - set(fail_ticket_ids)]
-        try:
+            need_delete_ticket_ids = get_deleted_tickets_ids(seadb_api, project_uuid)
+            if not need_delete_ticket_ids:
+                return Response({'success': True}, status=status.HTTP_200_OK)
             delete_ticket_comments_by_ids(seadb_api, project_uuid, need_delete_ticket_ids)
             seadb_api.delete_rows(project_uuid, 'tickets', need_delete_ticket_ids)
         except Exception as e:
@@ -1302,8 +1329,5 @@ class TicketTrashAPIView(APIView):
             error_msg = 'Internal Server Error'
             return api_error(status.HTTP_500_INTERNAL_SERVER_ERROR, error_msg)
 
-        return Response({
-            'success': need_delete_ticket_ids,
-            'failed': fail_ticket_ids
-        })
+        return Response({'success': True}, status=status.HTTP_200_OK)
